added clique
