+++ qiskit_optimization/applications/ising/max_cut.py
@@ -7,8 +7,8 @@ from .graph_application import GraphApplication
 
 class Maxcut(GraphApplication):
 
-    def __init__(self, g):
-        super().__init__(g)
+    def __init__(self, graph):
+        super().__init__(graph)
 
     def to_quadratic_program(self):
         mdl = Model(name='maxcut')
