+++ qiskit_optimization/applications/ising/tsp.py
@@ -10,8 +10,8 @@ from qiskit_optimization.problems.quadratic_program import QuadraticProgram
 
 class TSP(GraphApplication):
 
-    def __init__(self, g):
-        super().__init__(g)
+    def __init__(self, graph):
+        super().__init__(graph)
 
     def to_quadratic_program(self):
         mdl = Model(name='tsp')
